% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive-package.R
\docType{package}
\name{pophive-package}
\alias{pophive}
\alias{pophive-package}
\title{pophive: PopHIVE: Population Health Information and Visualization Exchange}
\description{
Collection and use of population health data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://DISSC-yale.github.io/pophive_demo}
  \item \url{https://github.com/DISSC-yale/pophive_demo}
  \item Report bugs at \url{https://github.com/DISSC-yale/pophive_demo/issues}
}

}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Other contributors:
\itemize{
  \item Yale School of Public Health [funder, copyright holder]
}

}
\keyword{internal}
