% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_add_source.R
\name{pophive_add_source}
\alias{pophive_add_source}
\title{Adds a source project structure}
\usage{
pophive_add_source(name, base_dir = "data")
}
\arguments{
\item{name}{Name of the source.}

\item{base_dir}{Path to the directory containing sources.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data source project, used to collect and prepare data from a new source.
}
\section{Project}{


Within a source project, there are two files to edits:
\itemize{
  \item \strong{\code{ingest.R}}: This is the primary script, which is automatically rerun.
    It should store raw data and resources in \code{raw/} where possible,
    then use what's in \code{raw/} to produce standard-format files in \code{standard/}.
    This file is sourced from its location during processing, so any system paths
    must be relative to itself.
  \item \strong{\code{measure_info.json}}: This is where you can record information
    about the variables included in the standardized data files.
    See \code{\link[community]{data_measure_info}}.
}
}

\examples{
data_source_dir <- tempdir()
pophive_add_source("source_name", data_source_dir)
list.files(paste0(data_source_dir, "/source_name"))

}
