% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_check_sources.R
\name{pophive_check_sources}
\alias{pophive_check_sources}
\title{Check Data Sources}
\usage{
pophive_check_sources(
  names = list.dirs("data", recursive = FALSE, full.names = FALSE),
  source_dir = "data",
  verbose = TRUE
)
}
\arguments{
\item{names}{Name or names of source projects.}

\item{source_dir}{Path to the directory containing the source projects.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with an entry for each source, containing a character vector
  including any issue codes:
\itemize{
  \item \code{not_compressed}: The file does not appear to be compressed.
  \item \code{cant_read}: Failed to read the file in.
  \item \code{geography_missing}: File does not contain a \code{geography} column.
  \item \code{geography_nas}: The file's \code{geography} column contains NAs.
  \item \code{time_missing}: File does not contain a \code{time} column.
  \item \code{time_nas}: The file's \code{time} column contains NAs.
  \item \code{missing_info: {column_name}}: The file's indicated column does not have
    a matching entry in \code{measure_info.json}.
}
}
\description{
Check the data files and measure info of source projects.
}
\examples{
\dontrun{
  pophive_check_sources("gtrends")
}
}
