% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_download_cdc.R
\name{pophive_download_cdc}
\alias{pophive_download_cdc}
\title{Download Data from the CDC}
\usage{
pophive_download_cdc(id, out_dir = "raw", verbose = TRUE)
}
\arguments{
\item{id}{ID of the resource (e.g., \code{ijqb-a7ye}).}

\item{out_dir}{Directory in which to save the metadata and data files.}

\item{verbose}{Logical; if \code{FALSE}, will not display status messages.}
}
\value{
Nothing; downloads files (\code{<id>.json} and \code{<id>.csv.xz}) to \code{out_dir}
}
\description{
Download data and metadata from the Centers for Disease Control and Prevention (CDC).
}
\section{\code{data.cdc.gov} URLs}{


For each resource ID, there are 3 relevant CDC URLs:
\itemize{
  \item \strong{\code{resource/<id>}}: This redirects to the resource's main page,
    with displayed metadata and a data preview
    (e.g., \code{\href{https://data.cdc.gov/resource/ijqb-a7ye}{data.cdc.gov/resource/ijqb-a7ye}}).
  \item \strong{\code{api/views/<id>}}: This is a direct link to the underlying
    JSON metadata (e.g., \code{\href{https://data.cdc.gov/api/views/ijqb-a7ye}{data.cdc.gov/api/views/ijqb-a7ye}}).
  \item \strong{\code{api/views/<id>/rows.csv}}: This is a direct link to the full
    CSV dataset (e.g., \code{\href{https://data.cdc.gov/api/views/ijqb-a7ye/rows.csv}{data.cdc.gov/api/views/ijqb-a7ye/rows.csv}}).
}
}

\examples{
\dontrun{
  pophive_download_cdc("ijqb-a7ye")
}
}
