% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_get_file.R
\name{pophive_get_file}
\alias{pophive_get_file}
\title{Retrieve A Data File}
\usage{
pophive_get_file(path, date = NULL, commit_hash = NULL, versions = FALSE)
}
\arguments{
\item{path}{Path to the file.}

\item{date}{Date of the version to load; A \code{Date}, or \code{character} in the format
\code{YYYY-MM-DD}. Will match to the nearest version.}

\item{commit_hash}{SHA signature of the committed version;
can be the first 6 or so characters. Ignored if \code{date} is provided.}

\item{versions}{Logical; if \code{TRUE}, will return a list of available version,
rather than a}
}
\value{
If \code{versions} is \code{TRUE}, a \code{data.frame} with columns for
  the \code{hash}, \code{author}, \code{date}, and \code{message} of each commit.
  Otherwise, the path to a temporary file, if one was extracted.
}
\description{
Load a data file from a source data project, or list versions of the file.
}
\examples{
path <- "../../data/wastewater/raw/flua.csv.xz"
if (file.exists(path)) {
  # list versions
  versions <- pophive_get_file(path, versions = TRUE)
  print(versions[, c("date", "hash")])

  # extract a version to a temporary file
  temp_path <- pophive_get_file(path, "2025-05")
  basename(temp_path)
}

}
