% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_load_census.R
\name{pophive_load_census}
\alias{pophive_load_census}
\title{Download Census Population Data}
\usage{
pophive_load_census(
  year = 2021,
  out_dir = NULL,
  state_only = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Data year.}

\item{out_dir}{Directory to download the file to.}

\item{state_only}{Logical; if \code{TRUE}, will only load state data.
Will still download county data.}

\item{overwrite}{Logical; if \code{TRUE}, will re-download and overwrite existing data.}

\item{verbose}{Logical; if \code{FALSE}, will not display status messages.}
}
\value{
A \code{data.frame} including \code{GEOID} and \code{region_name}
  for states and counties, along with their population, in total and within
  age brackets.
}
\description{
Download American Community Survey population data from the U.S. Census Bureau.
}
\examples{
if (file.exists("../../resources/census_population_2021.csv.xz")) {
  pophive_load_census(2021, "../../resources")[1:10, ]
}
}
