% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_process.R
\name{pophive_process}
\alias{pophive_process}
\title{Run Data Sources}
\usage{
pophive_process(name = NULL, source_dir = "data", ingest = TRUE)
}
\arguments{
\item{name}{Name of a source project to process. Will}

\item{source_dir}{Path to the directory containing source projects.}

\item{ingest}{Logical; if \code{FALSE}, will re-process standardized data without running
ingestion scripts.}
}
\value{
A list with processing results:
\itemize{
  \item \code{timings}: How many seconds the ingestion script took to run.
  \item \code{logs}: The captured output of the ingestion script.
}
Each entry has an entry for each source.

A `datapackage.json` file is also created / update in each source's `standard` directory.
}
\description{
Optionally run the ingestion script for each data source, and collect metadata.
}
\examples{
\dontrun{
  # run from a directory containing a `data` directory containing the source
  pophive_process("source_name")

  # run without executing the ingestion script
  pophive_process("source_name", ingest = FALSE)
}
}
