% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_read_epic.R
\name{pophive_read_epic}
\alias{pophive_read_epic}
\title{Read Epic Cosmos Data}
\usage{
pophive_read_epic(path, path_root = ".")
}
\arguments{
\item{path}{Path to the file.}

\item{path_root}{Directory containing \code{path}, if it is not full.}
}
\value{
A list with \code{data.frame} entries for \code{metadata} and \code{data}.
}
\description{
Read in metadata and data from an Epic Cosmos file.
}
\examples{
# write an example file
path <- tempfile(fileext = ".csv")
raw_lines <- c(
  "metadata field,metadata value,",
  ",,",
  ",Measures,Value Name",
  "Year,Measure 1,",
  "2020,m1,1",
  ",m2,2",
  "2021,m1,3",
  ",m2,4"
)
writeLines(raw_lines, path)

# read it in
pophive_read_epic(basename(path), dirname(path))

}
