% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_source_process.R
\name{pophive_source_process}
\alias{pophive_source_process}
\title{Interact with a Source Process File}
\usage{
pophive_source_process(path = "process.json", updated = NULL)
}
\arguments{
\item{path}{Path to the process JSON file.}

\item{updated}{An update version of the process definition. If specified, will
write this as the new process file, rather than reading any existing file.}
}
\value{
The process definition of the source project.
}
\description{
Read or update the current source process file.
}
\examples{
epic_process_file <- "../../epic/process.json"
if (file.exists(epic_process_file)) {
  pophive_source_process(path = epic_process_file)
}
}
