% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophive_standardize_epic.R
\name{pophive_standardize_epic}
\alias{pophive_standardize_epic}
\title{Standardize Epic Data}
\usage{
pophive_standardize_epic(raw_data)
}
\arguments{
\item{raw_data}{Raw Epic data, such as returned from \link{pophive_read_epic}.}
}
\value{
A standardized form of \code{data}.
}
\description{
Standardize a raw Epic data table.
}
\section{Standardization}{

\itemize{
  \item Collapse location columns (\code{state} or \code{county}) to a single
    \code{geography} column, and region names to IDs.
  \item Collapse time columns (\code{year}, \code{month}, or \code{week}) to a single
    \code{time} column, and clean up value formatting.
  \item Drop rows with no values across value columns.
}
}

\examples{
\dontrun{
  raw_data <- pophive_read_epic("data/epic/raw/flu.csv.xz")
  standard_data <- pophive_process_epic_raw(raw_data)
}

}
